vendor: update wireguard-go to v0.0.0-20210307
