+++ wireguard-go-bridge/go.mod
@@ -1,8 +1,8 @@
-module golang.zx2c4.com/wireguard/ios
+module golang.zx2c4.com/wireguard/apple
 
-go 1.15
+go 1.16
 
 require (
-	golang.org/x/sys v0.0.0-20201223074533-0d417f636930
-	golang.zx2c4.com/wireguard v0.0.20201119-0.20201223215156-09728dc6b340
+	golang.org/x/sys v0.0.0-20210308170721-88b6017d0656
+	golang.zx2c4.com/wireguard v0.0.0-20210307162820-f4695db51c39
 )
