+++ wireguard-go-bridge/wireguard.h
@@ -1,6 +1,6 @@
 /* SPDX-License-Identifier: GPL-2.0
  *
- * Copyright (C) 2018-2019 WireGuard LLC. All Rights Reserved.
+ * Copyright (C) 2018-2020 WireGuard LLC. All Rights Reserved.
  */
 
 #ifndef WIREGUARD_H
@@ -10,15 +10,14 @@
 #include <stdint.h>
 #include <stdbool.h>
 
-typedef struct { const char *p; size_t n; } gostring_t;
-typedef void(*logger_fn_t)(int level, const char *msg);
-extern void wgEnableRoaming(bool enabled);
-extern void wgSetLogger(logger_fn_t logger_fn);
+typedef void(*logger_fn_t)(void *context, int level, const char *msg);
+extern void wgSetLogger(void *context, logger_fn_t logger_fn);
 extern int wgTurnOn(const char *settings, int32_t tun_fd);
 extern void wgTurnOff(int handle);
 extern int64_t wgSetConfig(int handle, const char *settings);
 extern char *wgGetConfig(int handle);
 extern void wgBumpSockets(int handle);
+extern void wgDisableSomeRoamingForBrokenMobileSemantics(int handle);
 extern const char *wgVersion();
 
 #endif
